backend/requirements.txt
Django>=5.0,<6.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.5
django-cors-headers>=4.0
django-filter>=24.0
gunicorn>=22.0
